% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{augmented_input_requirement}
\alias{augmented_input_requirement}
\title{Augmented input requirement}
\usage{
augmented_input_requirement(X, w, c, d)
}
\arguments{
\item{X}{transaction matrix}

\item{w}{wage vector}

\item{c}{household consumption vector}

\item{d}{final demand vector}
}
\description{
Augmented input requirement
}
\examples{
set.seed(200100)
X <- matrix(rnorm(100), nrow = 10)
w <- rnorm(10)
c <- rnorm(10)
d <- rnorm(10)
augmented_input_requirement(X,w,c,d)
}
