% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{backward_linkage_cv}
\alias{backward_linkage_cv}
\title{Backward linkage coefficient of variation}
\usage{
backward_linkage_cv(L)
}
\arguments{
\item{L}{Leontief inverse matrix}
}
\description{
Backward linkage coefficient of variation
}
