% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputoutput-package.R
\docType{data}
\name{demand_matrix_12x12}
\alias{demand_matrix_12x12}
\title{Demand matrix (rev. 2008)
This matrix contains the production of the chilean economy divided into
12 industries. The intermediate and total demand is also included to support
the vignettes. The measuring unit is CLP million of the year 2008.}
\format{
A matrix with 12 rows and 14 columns
}
\usage{
demand_matrix_12x12
}
\description{
Demand matrix (rev. 2008)
This matrix contains the production of the chilean economy divided into
12 industries. The intermediate and total demand is also included to support
the vignettes. The measuring unit is CLP million of the year 2008.
}
\references{
\url{https://si3.bcentral.cl/estadisticas/Principal1/Excel/CCNN/cdr/excel.html}
}
\author{
Central Bank of Chile
}
\keyword{data}
