% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leontief-package.R
\docType{data}
\name{employment_matrix}
\alias{employment_matrix}
\title{Employment matrix (2013 data)
This matrix contains the employed people by industry and the employment coefficient
that is the number of people divided by the total final demand from the wage and
demand matrix.}
\format{
A matrix with 12 rows and 2 columns
}
\usage{
wage_demand_matrix
}
\description{
Employment matrix (2013 data)
This matrix contains the employed people by industry and the employment coefficient
that is the number of people divided by the total final demand from the wage and
demand matrix.
}
\references{
\url{http://revistas.ubiobio.cl/index.php/HHEE/article/download/3441/3473/}
}
\author{
University of Bio-Bio, based on data from the National Bureau of Statistics
}
\keyword{data}
