% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{employment_number}
\alias{employment_number}
\title{Employment number}
\usage{
employment_number(L, e, c)
}
\arguments{
\item{L}{Leontief inverse matrix}

\item{e}{employment coefficients vector}

\item{c}{change in final demand}
}
\description{
Employment number
}
