% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leontief-matrices.R
\name{equilibrium_output}
\alias{equilibrium_output}
\title{Equilibrium output}
\usage{
equilibrium_output(L, d)
}
\arguments{
\item{L}{Leontief inverse matrix}

\item{d}{final demand vector}
}
\description{
Equilibrium output
}
\examples{
set.seed(200100)
L <- matrix(rnorm(100), nrow = 10)
d <- rnorm(10)
equilibrium_output(L, d)
}
