% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{forward_linkage_cv}
\alias{forward_linkage_cv}
\title{Forward linkage coefficient of variation}
\usage{
forward_linkage_cv(L)
}
\arguments{
\item{L}{Leontief inverse matrix}
}
\description{
Forward linkage coefficient of variation
}
