% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{household_income_multiplier}
\alias{household_income_multiplier}
\title{Household income multiplier}
\usage{
household_income_multiplier(L, w)
}
\arguments{
\item{L}{Leontief inverse matrix}

\item{w}{wage vector}
}
\description{
Household income multiplier
}
