% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leontief-matrices.R
\name{income_multiplier}
\alias{income_multiplier}
\title{Income multiplier}
\usage{
income_multiplier(L, w)
}
\arguments{
\item{L}{Leontief inverse matrix}

\item{w}{wage vector}
}
\description{
Income multiplier
}
