% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{input_multiplier}
\alias{input_multiplier}
\title{Input multiplier}
\usage{
input_multiplier(G)
}
\arguments{
\item{G}{Ghosh inverse matrix}
}
\description{
Input multiplier
}
\examples{
set.seed(200100)
G <- matrix(rnorm(100), nrow = 10)
output_multiplier(G)
}
