% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leontief-matrices.R
\name{input_requirement}
\alias{input_requirement}
\title{Input requirement}
\usage{
input_requirement(X, d)
}
\arguments{
\item{X}{transaction matrix}

\item{d}{final demand vector}
}
\description{
Input requirement
}
\examples{
set.seed(200100)
X <- matrix(rnorm(100), nrow = 10)
d <- rnorm(10)
input_requirement(X, d)
}
