\name{inputoutput-package}
\alias{inputoutput-package}
\alias{inputoutput}
\docType{package}
\title{
\packageTitle{inputoutput}
}
\description{
\packageDescription{inputoutput}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{inputoutput}
\packageIndices{inputoutput}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{inputoutput}

Maintainer: \packageMaintainer{inputoutput}
}
\references{
  Venegas Morales, Jose. Una Matrix Insumo-Producto Inversa de la Economia Chilena. Banco Central de Chile, Chile, 1986.
}
\keyword{package}
