% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leontief-package.R
\docType{package}
\name{leontief-package}
\alias{leontief}
\alias{leontief-package}
\title{leontief: Input-Output Analysis}
\description{
An implementation of the Input-Output model developed by Wassily Leontief that represents the interdependencies between different sectors of a national economy or different regional economies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pachadotdev/leontief}
  \item Report bugs at \url{https://github.com/pachadotdev/leontief/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Other contributors:
\itemize{
  \item Victor Falkenheim [thesis advisor]
  \item  Central Bank of Chile [data contributor]
  \item  University of Bio-Bio [data contributor]
}

}
\keyword{internal}
