% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leontief-matrices.R
\name{multiplier_product_matrix}
\alias{multiplier_product_matrix}
\title{Multiplier product matrix}
\usage{
multiplier_product_matrix(L)
}
\arguments{
\item{L}{Leontief inverse matrix}
}
\description{
Multiplier product matrix
}
