% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leontief-matrices.R
\name{output_multiplier}
\alias{output_multiplier}
\title{Output multiplier}
\usage{
output_multiplier(L)
}
\arguments{
\item{L}{Leontief inverse matrix}
}
\description{
Output multiplier
}
\examples{
set.seed(200100)
L <- matrix(rnorm(100), nrow = 10)
output_multiplier(L)
}
