% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leontief-matrices.R
\name{power_dispersion_cv}
\alias{power_dispersion_cv}
\title{Power of dispersion coefficient of variation}
\usage{
power_dispersion_cv(L)
}
\arguments{
\item{L}{Leontief inverse matrix}
}
\description{
Power of dispersion coefficient of variation
}
