% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leontief-matrices.R
\name{sensitivity_dispersion}
\alias{sensitivity_dispersion}
\title{Sensitivity of dispersion coefficient of variation}
\usage{
sensitivity_dispersion(L)
}
\arguments{
\item{L}{Leontief inverse matrix}
}
\description{
Sensitivity of dispersion coefficient of variation
}
