% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leontief-matrices.R
\name{sensitivity_dispersion_cv}
\alias{sensitivity_dispersion_cv}
\title{Sensititivy of dispersion coefficient of variation}
\usage{
sensitivity_dispersion_cv(L)
}
\arguments{
\item{L}{Leontief inverse matrix}
}
\description{
Sensititivy of dispersion coefficient of variation
}
