% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leontief-package.R
\docType{data}
\name{transaction_matrix}
\alias{transaction_matrix}
\title{Transaction matrix (2013 data)
This matrix contains the production of the chilean economy divided into
12 industries. The measuring unit is CLP million of the year 2013}
\format{
A matrix with 12 rows and 12 columns
}
\usage{
transaction_matrix
}
\description{
Transaction matrix (2013 data)
This matrix contains the production of the chilean economy divided into
12 industries. The measuring unit is CLP million of the year 2013
}
\references{
\url{https://si3.bcentral.cl/estadisticas/Principal1/Excel/CCNN/cdr/excel.html}
}
\author{
Central Bank of Chile
}
\keyword{data}
