% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leontief-package.R
\docType{data}
\name{wage_demand_matrix}
\alias{wage_demand_matrix}
\title{Wage and demand matrix (2013 data)
This matrix contains the wage, intermediate demand and disaggregated final
demand of the  chilean economy divided into 12 industries. The final demand
is given by components (household consumption, government consumption, etc.)
and aggregated. The measuring unit is CLP million of the year 2013.}
\format{
A matrix with 12 rows and 9 columns
}
\usage{
wage_demand_matrix
}
\description{
Wage and demand matrix (2013 data)
This matrix contains the wage, intermediate demand and disaggregated final
demand of the  chilean economy divided into 12 industries. The final demand
is given by components (household consumption, government consumption, etc.)
and aggregated. The measuring unit is CLP million of the year 2013.
}
\references{
\url{https://si3.bcentral.cl/estadisticas/Principal1/Excel/CCNN/cdr/excel.html}
}
\author{
Central Bank of Chile
}
\keyword{data}
