% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leontief-matrices.R
\name{employment_multiplier}
\alias{employment_multiplier}
\title{Employment multiplier}
\usage{
employment_multiplier(L, e)
}
\arguments{
\item{L}{Leontief inverse matrix}

\item{e}{employment coefficients vector}
}
\description{
Employment multiplier
}
