% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ghosh_inverse}
\alias{ghosh_inverse}
\title{Ghosh inverse matrix}
\usage{
ghosh_inverse(B)
}
\arguments{
\item{B}{output allocation matrix}
}
\description{
Ghosh inverse matrix
}
\examples{
set.seed(200100)
B <- matrix(rnorm(100), nrow = 10)
leontief_inverse(B)
}
