% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{input_requirement_matrix}
\alias{input_requirement_matrix}
\title{Input requirement matrix}
\usage{
input_requirement_matrix(X, d)
}
\arguments{
\item{X}{input-output matrix}

\item{d}{final demand vector}
}
\description{
Input requirement matrix
}
\examples{
set.seed(200100)
X <- matrix(rnorm(100), nrow = 10)
d <- rnorm(10)
input_requirement_matrix(X,d)
}
