% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leontief-matrices.R
\name{leontief_inverse}
\alias{leontief_inverse}
\title{Leontief inverse}
\usage{
leontief_inverse(A)
}
\arguments{
\item{A}{input requirement matrix}
}
\description{
Leontief inverse
}
\examples{
set.seed(200100)
A <- matrix(rnorm(100), nrow = 10)
leontief_inverse(A)
}
