% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{output_allocation_matrix}
\alias{output_allocation_matrix}
\title{Output allocation matrix}
\usage{
output_allocation_matrix(X, d)
}
\arguments{
\item{X}{input-output matrix}

\item{d}{final demand vector}
}
\description{
Output allocation matrix
}
\examples{
set.seed(200100)
X <- matrix(rnorm(100), nrow = 10)
d <- rnorm(10)
output_allocation_matrix(X,d)
}
